implemented code suggestions from giovp
