+++ src/spatialdata/models/models.py
@@ -35,14 +35,13 @@ from xarray_schema.dataarray import DataArraySchema
 
 from spatialdata._logging import logger
 from spatialdata._types import ArrayLike
-from spatialdata.models import C, X, Y, Z
+from spatialdata.models import C, X, Y, Z, get_axis_names
 from spatialdata.models._utils import (
     DEFAULT_COORDINATE_SYSTEM,
     TRANSFORM_KEY,
     MappingToCoordinateSystem_t,
     SpatialElement,
     _validate_mapping_to_coordinate_system_type,
-    get_axis_names,
 )
 from spatialdata.transformations._utils import (
     _get_transformations,
