+++ tests/test_io/test_utils.py
@@ -5,9 +5,7 @@ import dask.dataframe as dd
 
 from spatialdata import read_zarr
 from spatialdata._io._utils import get_backing_files
-from spatialdata.element_utils.element_utils import (
-    multiscale_spatial_image_from_data_tree,
-)
+from spatialdata._utils import multiscale_spatial_image_from_data_tree
 
 
 def test_backing_files_points(points):
